guava/backend/shared/requirements.txt
# Shared requirements for all microservices
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-filter>=23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pika>=1.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0


